test: 레포지토리 있는 엔티티는 이동
